Improve app model
