+++ src/lib/app/app.ts
@@ -1,0 +1,11 @@
+import { Organization } from './organization';
+
+class Application {
+	organization: Organization;
+
+	constructor() {
+		this.organization = new Organization();
+	}
+}
+
+export const app = new Application();
