+++ src/lib/app/backlog.ts
@@ -1,0 +1,7 @@
+import type { Feature } from './feature';
+
+export class Backlog {
+	features: Record<string, Feature> = {};
+
+	constructor() {}
+}
