+++ src/lib/app/storable.ts
@@ -1,0 +1,5 @@
+import type { RedisStore } from '$lib/stores/redis-rune-store.svelte';
+
+export abstract class Storable {
+	stores: Record<string, RedisStore> = {};
+}
