+++ src/lib/server/redis.ts
@@ -1,12 +1,9 @@
 // import Redis from 'ioredis';
 import RedisMock from 'ioredis-mock';
-// import { createRedisStore } from '$lib/stores/redis-store';
 import { RedisStore } from '$lib/stores/redis-rune-store.svelte';
 
 export const redis = new RedisMock(6380); // new Redis(process.env.REDIS_URL);
-export const pubsub = redis; //.duplicate();
-
-//export const messageStore = createRedisStore(pubsub, 'myChannel');
+export const pubsub = redis;
 
 export const createStore = (name: string = '', type: string) =>
 	new RedisStore(pubsub, `${name}@${type}`);
