+++ src/routes/projects/+page.server.ts
@@ -1,0 +1,9 @@
+export const load = async () => {
+	return {
+		projects: [
+			{
+				name: 'My project'
+			}
+		]
+	};
+};
